requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
pydantic==1.9.0
pytz==2021.3
pyvalid==1.0.4
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
typing_extensions==4.0.1
